Conservative refactoring Plot
* Plot and ChartSeries classes are both created in `shoes/types`
* Shall we move plot directory to `shoes/types` ?
* Plot is huge. Handle with care.
* Shoes3/shoes3#333
