+++ shoes/plot/plot.h
@@ -152,6 +152,7 @@ extern void shoes_plot_draw_label(cairo_t *, shoes_plot *, int, int, char*, int)
 extern void shoes_plot_draw_nub(cairo_t *, shoes_plot *, double, double, int, int);
 extern int shoes_plot_util_quadrant(double angle);
 
-
+// TODO: refactor into shoes/types?
+extern VALUE cPlot, cChartSeries;
 
 #endif
