+++ shoes/ruby.c
@@ -14,7 +14,6 @@
 #include <math.h>
 
 VALUE cShoes, cApp, cDialog, cTypes, cShoesWindow, cMouse, cCanvas, cFlow, cStack, cMask, cWidget, cShape, cImage, cPattern, cBorder, cBackground, cTextBlock, cPara, cBanner, cTitle, cSubtitle, cTagline, cCaption, cInscription, cTextClass, cNative, cCheck, cRadio, cProgress, cColor, cDownload, cResponse, ssNestSlot;
-VALUE cPlot, cChartSeries;
 VALUE eImageError, eInvMode, eNotImpl;
 VALUE reHEX_SOURCE, reHEX3_SOURCE, reRGB_SOURCE, reRGBA_SOURCE, reGRAY_SOURCE, reGRAYA_SOURCE, reLF;
 VALUE symAltQuest, symAltSlash, symAltDot, symAltEqual, symAltSemiColon;
@@ -2395,10 +2394,6 @@ PLACE_COMMON(textblock)
 CLASS_COMMON2(textblock)
 REPLACE_COMMON(textblock)
 
-// The next two macros are very important for new widget writers.
-CLASS_COMMON2(plot)
-TRANS_COMMON(plot, 1);
-
 VALUE shoes_textblock_style_m(int argc, VALUE *argv, VALUE self) {
     GET_STRUCT(textblock, self_t);
     VALUE obj = shoes_textblock_style(argc, argv, self);
@@ -2965,69 +2960,6 @@ void shoes_ruby_init() {
     rb_define_method(cImage, "release", CASTHOOK(shoes_image_release), -1);
     rb_define_method(cImage, "hover", CASTHOOK(shoes_image_hover), -1);
     rb_define_method(cImage, "leave", CASTHOOK(shoes_image_leave), -1);
-
-    cChartSeries = rb_define_class_under(cTypes, "chart_series", rb_cObject); // 3.3.2
-    rb_define_alloc_func(cChartSeries, shoes_chart_series_alloc);
-    //  simple getters/setters
-    rb_define_method(cChartSeries, "values", CASTHOOK(shoes_chart_series_values), 0);
-    rb_define_method(cChartSeries, "labels", CASTHOOK(shoes_chart_series_labels), 0);
-    rb_define_method(cChartSeries, "min", CASTHOOK(shoes_chart_series_min), 0);
-    rb_define_method(cChartSeries, "min=", CASTHOOK(shoes_chart_series_min_set), 1);
-    rb_define_method(cChartSeries, "max", CASTHOOK(shoes_chart_series_max), 0);
-    rb_define_method(cChartSeries, "max=", CASTHOOK(shoes_chart_series_max_set), 1);
-    rb_define_method(cChartSeries, "name", CASTHOOK(shoes_chart_series_name), 0);
-    rb_define_method(cChartSeries, "desc", CASTHOOK(shoes_chart_series_desc), 0);
-    rb_define_method(cChartSeries, "desc=", CASTHOOK(shoes_chart_series_desc_set), 1);
-    rb_define_method(cChartSeries, "color", CASTHOOK(shoes_chart_series_color), 0);
-    rb_define_method(cChartSeries, "color=", CASTHOOK(shoes_chart_series_color_set), 1);
-    rb_define_method(cChartSeries, "strokewidth", CASTHOOK(shoes_chart_series_strokewidth), 0);
-    rb_define_method(cChartSeries, "strokewidth=", CASTHOOK(shoes_chart_series_strokewidth_set), 1);
-    rb_define_method(cChartSeries, "points", CASTHOOK(shoes_chart_series_points), 0);
-    rb_define_method(cChartSeries, "points=", CASTHOOK(shoes_chart_series_points_set), 1);
-    // more complcated methods
-    rb_define_method(cChartSeries, "at", CASTHOOK(shoes_chart_series_get), 1);
-    rb_define_method(cChartSeries, "get", CASTHOOK(shoes_chart_series_get), 1);
-    rb_define_method(cChartSeries, "set", CASTHOOK(shoes_chart_series_set), 2);
-
-    cPlot   = rb_define_class_under(cTypes, "Plot", rb_cObject); // 3.3.2
-    rb_define_alloc_func(cPlot, shoes_plot_alloc);
-    // methods unique to plot
-    rb_define_method(cPlot, "add", CASTHOOK(shoes_plot_add), 1);
-    rb_define_method(cPlot, "redraw_to", CASTHOOK(shoes_plot_redraw_to), 1);
-    rb_define_method(cPlot, "delete", CASTHOOK(shoes_plot_delete), 1);
-    rb_define_method(cPlot, "id",  CASTHOOK(shoes_plot_find_name), 1);
-    rb_define_method(cPlot, "count", CASTHOOK(shoes_plot_get_count), 0);
-    rb_define_method(cPlot, "first", CASTHOOK(shoes_plot_get_first), 0);
-    rb_define_method(cPlot, "set_first", CASTHOOK(shoes_plot_set_first), 1);
-    rb_define_method(cPlot, "last", CASTHOOK(shoes_plot_get_last), 0);
-    rb_define_method(cPlot, "set_last", CASTHOOK(shoes_plot_set_last), 1);
-    rb_define_method(cPlot, "zoom", CASTHOOK(shoes_plot_zoom), 2);
-    rb_define_method(cPlot, "save_as", CASTHOOK(shoes_plot_save_as), -1);
-    rb_define_method(cPlot, "near_x", CASTHOOK(shoes_plot_near), 1);
-    // methods commom to many Shoes widgets
-    rb_define_method(cPlot, "draw", CASTHOOK(shoes_plot_draw), 2);
-    rb_define_method(cPlot, "remove", CASTHOOK(shoes_plot_remove), 0);
-    rb_define_method(cPlot, "parent", CASTHOOK(shoes_plot_get_parent), 0);
-    rb_define_method(cPlot, "style", CASTHOOK(shoes_plot_style), -1);
-    rb_define_method(cPlot, "move", CASTHOOK(shoes_plot_move), 2);
-    rb_define_method(cPlot, "displace", CASTHOOK(shoes_plot_displace), 2);
-    rb_define_method(cPlot, "hide", CASTHOOK(shoes_plot_hide), 0);
-    rb_define_method(cPlot, "show", CASTHOOK(shoes_plot_show), 0);
-    rb_define_method(cPlot, "toggle", CASTHOOK(shoes_plot_toggle), 0);
-    rb_define_method(cPlot, "hidden?", CASTHOOK(shoes_plot_is_hidden), 0);
-    rb_define_method(cPlot, "click", CASTHOOK(shoes_plot_click), -1);
-    rb_define_method(cPlot, "released", CASTHOOK(shoes_plot_release), -1);
-    rb_define_method(cPlot, "hover", CASTHOOK(shoes_plot_hover), -1);
-    rb_define_method(cPlot, "leave", CASTHOOK(shoes_plot_leave), -1);
-    rb_define_method(cPlot, "top", CASTHOOK(shoes_plot_get_actual_top), 0);
-    rb_define_method(cPlot, "left", CASTHOOK(shoes_plot_get_actual_left), 0);
-    rb_define_method(cPlot, "width", CASTHOOK(shoes_plot_get_actual_width), 0);
-    rb_define_method(cPlot, "height", CASTHOOK(shoes_plot_get_actual_height), 0);
-    rb_define_method(cPlot, "transform", CASTHOOK(shoes_plot_transform), 1);
-    rb_define_method(cPlot, "translate", CASTHOOK(shoes_plot_translate), 2);
-    rb_define_method(cPlot, "rotate", CASTHOOK(shoes_plot_rotate), 1);
-    rb_define_method(cPlot, "scale", CASTHOOK(shoes_plot_scale), -1);
-    rb_define_method(cPlot, "skew", CASTHOOK(shoes_plot_skew), -1);
 
     cPattern = rb_define_class_under(cTypes, "Pattern", rb_cObject);
     rb_define_alloc_func(cPattern, shoes_pattern_alloc);
