+++ shoes/types/types.h
@@ -4,6 +4,7 @@
 #include "shoes/types/edit_line.h"
 #include "shoes/types/effect.h"
 #include "shoes/types/list_box.h"
+#include "shoes/types/plot.h"
 #include "shoes/types/progress.h"
 #include "shoes/types/slider.h"
 #include "shoes/types/spinner.h"
@@ -22,6 +23,7 @@
 	shoes_edit_line_init(); \
 	shoes_effect_init(); \
 	shoes_list_box_init(); \
+	shoes_plot_init(); \
 	shoes_progress_init(); \
 	shoes_slider_init(); \
 	shoes_spinner_init(); \
